[9/9] feat(ui): remove redundant header chrome, delegate background to host (#80)
