--- src/components/llm-test-runner/header/llm-test-runner-header.css
@@ -1,96 +1,0 @@
-/* Test Runner Header Block */
-.test-runner-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: var(--spacing-4);
-  padding: var(--spacing-3) var(--spacing-5);
-  background: var(--background);
-  color: var(--foreground);
-  border-bottom: var(--border-width) solid var(--border);
-  position: sticky;
-  top: var(--llmtr-sticky-top, 0px);
-  z-index: var(--z-sticky);
-  box-shadow: var(--shadow-sm);
-}
-
-/* Header Elements */
-.test-runner-header__left,
-.test-runner-header__right {
-  display: flex;
-  gap: var(--spacing-3);
-  align-items: center;
-}
-
-.test-runner-header__center {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* API Status Element */
-.test-runner-header__api-status {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-2);
-}
-
-.test-runner-header__api-status-text {
-  color: var(--success);
-  font-weight: var(--font-weight-medium);
-  font-size: var(--font-size-sm);
-  line-height: var(--line-height-normal);
-}
-
-/* Modifier: Hidden */
-.test-runner-header--hidden {
-  display: none;
-}
-
-/* "Show summary" checkbox toggle */
-.test-runner-header__toggle {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--spacing-2);
-  cursor: pointer;
-  user-select: none;
-  padding: var(--spacing-1) var(--spacing-2);
-  border-radius: var(--radius-md);
-  color: var(--muted-foreground);
-  transition: color 150ms ease, background-color 150ms ease;
-}
-
-.test-runner-header__toggle:hover {
-  color: var(--foreground);
-  background: var(--muted);
-}
-
-.test-runner-header__toggle-input {
-  width: 16px;
-  height: 16px;
-  margin: 0;
-  cursor: pointer;
-  accent-color: var(--primary);
-}
-
-.test-runner-header__toggle-label {
-  font-size: var(--font-size-sm);
-  font-weight: var(--font-weight-medium);
-  letter-spacing: var(--letter-spacing-snug);
-}
-
-/* Responsive Design */
-@media (max-width: 768px) /* var(--breakpoint-md) */ {
-  .test-runner-header {
-    flex-direction: column;
-    gap: var(--spacing-4);
-    padding: var(--spacing-4);
-  }
-
-  .test-runner-header__left,
-  .test-runner-header__right {
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-}
