--- src/components/llm-test-runner/header/llm-test-runner-header.tsx
@@ -1,157 +1,0 @@
-import { h, FunctionalComponent } from '@stencil/core';
-import { Button } from '../../../lib/ui/button/index';
-import {
-  UploadIcon,
-  DownloadIcon,
-  FileTextIcon,
-  SaveIcon,
-  SlidersIcon,
-  PlayIcon,
-  SpinnerIcon,
-  PlusIcon,
-} from '../../../lib/ui/icons/icons';
-
-export interface LLMTestRunnerHeaderProps {
-  isExportingTestSuite: boolean;
-  isExportingTestResults: boolean;
-  isRunningAll: boolean;
-  canRunAny: boolean;
-  useSave?: boolean;
-  isSaving?: boolean;
-  usePromptEditor?: boolean;
-  showSummary: boolean;
-  onImport: (file: File) => void;
-  onExportSuite: () => void;
-  onExportResults: () => void;
-  onRunAll: () => void;
-  onSave?: () => void;
-  onAddTestCase: () => void;
-  onToggleSummary: (show: boolean) => void;
-}
-
-export const LLMTestRunnerHeader: FunctionalComponent<
-  LLMTestRunnerHeaderProps
-> = ({
-  isExportingTestSuite,
-  isExportingTestResults,
-  isRunningAll,
-  canRunAny,
-  useSave = false,
-  isSaving = false,
-  usePromptEditor = false,
-  showSummary,
-  onImport,
-  onExportSuite,
-  onExportResults,
-  onRunAll,
-  onSave,
-  onAddTestCase,
-  onToggleSummary,
-}) => {
-  let fileInputRef: HTMLInputElement;
-
-  const handleFileSelect = () => {
-    fileInputRef?.click();
-  };
-
-  const handleFileChange = (event: Event) => {
-    const target = event.target as HTMLInputElement;
-    const file = target.files?.[0];
-    target.value = ''; // Clear for re-upload
-    if (file) {
-      onImport(file);
-    }
-  };
-
-  return (
-    <header class="test-runner-header">
-      <div class="test-runner-header__left">
-        <input
-          class="test-runner-header--hidden"
-          type="file"
-          ref={el => (fileInputRef = el as HTMLInputElement)}
-          onChange={handleFileChange}
-          accept=".json,application/json"
-        />
-        <Button
-          variant="outline"
-          size="md"
-          onClick={handleFileSelect}
-          icon={<UploadIcon />}
-        >
-          Import suite
-        </Button>
-        <Button
-          variant="outline"
-          size="md"
-          onClick={onExportSuite}
-          disabled={isExportingTestSuite}
-          loading={isExportingTestSuite}
-          icon={isExportingTestSuite ? <SpinnerIcon /> : <DownloadIcon />}
-        >
-          {isExportingTestSuite ? 'Exporting…' : 'Export suite'}
-        </Button>
-        <label class="test-runner-header__toggle">
-          <input
-            type="checkbox"
-            class="test-runner-header__toggle-input"
-            checked={showSummary}
-            onChange={(e) =>
-              onToggleSummary((e.target as HTMLInputElement).checked)
-            }
-          />
-          <span class="test-runner-header__toggle-label">Show summary</span>
-        </label>
-      </div>
-
-      <div class="test-runner-header__right">
-        {usePromptEditor && (
-          <Button variant="outline" size="md" icon={<SlidersIcon />}>
-            Prompt editor
-          </Button>
-        )}
-        <Button
-          variant="outline"
-          size="md"
-          onClick={onAddTestCase}
-          icon={<PlusIcon />}
-        >
-          Add question
-        </Button>
-        <Button
-          variant="outline"
-          size="md"
-          onClick={onExportResults}
-          disabled={isExportingTestResults}
-          loading={isExportingTestResults}
-          icon={isExportingTestResults ? <SpinnerIcon /> : <FileTextIcon />}
-        >
-          {isExportingTestResults ? 'Exporting…' : 'Export results'}
-        </Button>
-        {useSave && (
-          <Button
-            variant="outline"
-            size="md"
-            onClick={onSave}
-            disabled={isSaving}
-            loading={isSaving}
-            icon={isSaving ? <SpinnerIcon /> : <SaveIcon />}
-          >
-            {isSaving ? 'Saving…' : 'Save'}
-          </Button>
-        )}
-        <Button
-          aria-label="Run all"
-          variant="primary"
-          size="md"
-          onClick={onRunAll}
-          disabled={isRunningAll || !canRunAny}
-          loading={isRunningAll}
-          icon={isRunningAll ? <SpinnerIcon /> : <PlayIcon />}
-        >
-          {isRunningAll ? 'Running…' : 'Run all'}
-        </Button>
-      </div>
-    </header>
-  );
-};
