+++ src/components/llm-test-runner/llm-test-runner.css
@@ -5,7 +5,6 @@
   letter-spacing: var(--letter-spacing-snug);
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: var(--background);
   color: var(--foreground);
 }
 
@@ -25,9 +24,6 @@
   width: 100%;
   margin: 0;
   padding: 0;
-  background: var(--background);
-  border: var(--border-width) solid var(--border);
-  border-radius: var(--radius-xl);
   overflow-x: clip;
   min-height: 100%;
 }
