+++ src/components/llm-test-runner/llm-test-runner.tsx
@@ -44,14 +44,10 @@ import {
   validateExpectedOutcomeArrayWithExtractors,
   validateExpectedOutcomeSchema,
 } from '../../schemas/expected-outcome';
-import { LLMTestRunnerHeader } from './header/llm-test-runner-header';
-import {
-  LLMTestRunnerSummary,
-  computeSummaryStats,
-} from './summary/llm-test-runner-summary';
 import { LLMTestCases } from './test-cases/llm-test-cases';
 import { TestCasesToolbar } from './test-cases/test-cases-toolbar';
 import {
+  computeSummaryStats,
   filterTestCasesByStatus,
   filterTestCasesByQuery,
   type StatusFilter,
@@ -71,8 +67,6 @@ function nextFrame(): Promise<void> {
   styleUrls: [
     '../../styles/tokens.css',
     'llm-test-runner.css',
-    'header/llm-test-runner-header.css',
-    'summary/llm-test-runner-summary.css',
     'test-cases/llm-test-cases.css',
     'test-cases/test-cases-toolbar.css',
     'test-cases/llm-test-case-row.css',
@@ -119,7 +113,6 @@ export class LLMTestRunner {
   @State() isExportingTestSuite: boolean = false;
   @State() isExportingTestResults: boolean = false;
   @State() isSaving: boolean = false;
-  @State() showSummary: boolean = true;
   @State() activeFilter: StatusFilter = 'all';
   @State() searchQuery: string = '';
   @State() isSearchExpanded: boolean = false;
@@ -484,28 +477,6 @@ export class LLMTestRunner {
     const stats = computeSummaryStats(this.testCases);
     return (
       <div class="test-runner-container">
-        <LLMTestRunnerHeader
-          isExportingTestSuite={this.isExportingTestSuite}
-          isExportingTestResults={this.isExportingTestResults}
-          isRunningAll={this.isRunningAll}
-          canRunAny={this.testCases.some(tc =>
-            isTestCaseRunnable(tc),
-          )}
-          useSave={this.useSave}
-          isSaving={this.isSaving}
-          usePromptEditor={this.usePromptEditor}
-          showSummary={this.showSummary}
-          onImport={file => this.handleImport(file)}
-          onExportSuite={() => this.handleExportTestSuite()}
-          onExportResults={() => this.handleExportTestResults()}
-          onRunAll={() => this.runAllTests()}
-          onSave={() => this.handleSave()}
-          onAddTestCase={() => this.addNewTestCase()}
-          onToggleSummary={(show) => (this.showSummary = show)}
-        />
-        {this.showSummary && (
-          <LLMTestRunnerSummary testCases={this.testCases} />
-        )}
         <ErrorMessage message={this.error} onClear={() => (this.error = '')} />
         <div class="test-runner-container__content">
           <TestCasesToolbar
@@ -515,11 +486,22 @@ export class LLMTestRunner {
             passedCount={stats.passed}
             activeFilter={this.activeFilter}
             isExportingTestSuite={this.isExportingTestSuite}
+            isExportingTestResults={this.isExportingTestResults}
+            isRunningAll={this.isRunningAll}
+            canRunAny={this.testCases.some(tc =>
+              isTestCaseRunnable(tc),
+            )}
+            useSave={this.useSave}
+            isSaving={this.isSaving}
+            usePromptEditor={this.usePromptEditor}
             searchQuery={this.searchQuery}
             isSearchExpanded={this.isSearchExpanded}
             onAddTestCase={() => this.addNewTestCase()}
             onImport={file => this.handleImport(file)}
             onExportSuite={() => this.handleExportTestSuite()}
+            onExportResults={() => this.handleExportTestResults()}
+            onRunAll={() => this.runAllTests()}
+            onSave={() => this.handleSave()}
             onFilterChange={filter => (this.activeFilter = filter)}
             onSearchQueryChange={query => (this.searchQuery = query)}
             onSearchExpandedChange={expanded => (this.isSearchExpanded = expanded)}
