+++ src/components/llm-test-runner/test-cases/test-cases-toolbar.css
@@ -1,12 +1,15 @@
-/* Shares .test-cases's background + horizontal padding so the two sit on one
- * continuous tray with no seam. */
+/* No background of its own — sits on whatever the host page provides, same as
+ * the rest of the component (see llm-test-runner.css/:host). Sticky via
+ * --llmtr-sticky-top since it now holds Run all and the other primary actions. */
 .test-cases-toolbar {
   display: flex;
-  align-items: center;
+  align-items: start;
   justify-content: space-between;
   gap: var(--spacing-2);
   padding: var(--spacing-3) var(--spacing-5) var(--spacing-2);
-  background: var(--muted);
+  position: sticky;
+  top: var(--llmtr-sticky-top, 0px);
+  z-index: var(--z-sticky);
 }
 
 .test-cases-toolbar__left {
@@ -45,8 +48,8 @@
   border-left: var(--border-width) solid var(--border-strong);
 }
 
-/* --secondary resolves to this toolbar's --muted background in this token
- * set, which would make an inactive pill invisible — use --background instead. */
+/* --secondary resolves to the same value as --muted in this token set — a
+ * likely host-page background — which would make an inactive pill invisible. */
 .test-cases-toolbar__filter-pill {
   display: inline-flex;
   align-items: center;
@@ -83,12 +86,20 @@
   color: var(--primary-foreground);
 }
 
+/* Grouped as one flex child so justify-content: space-between pushes it right as a unit. */
+.test-cases-toolbar__right {
+  display: flex;
+  align-items: center;
+  gap: var(--spacing-2);
+  flex-shrink: 0;
+}
+
 .test-cases-toolbar__search {
   flex-shrink: 0;
 }
 
 .test-cases-toolbar__search-input {
-  width: 220px;
+  width: 160px;
   height: 32px;
   box-sizing: border-box;
   padding: 0 var(--spacing-3);
