+++ src/components/llm-test-runner/test-cases/test-cases-toolbar.tsx
@@ -1,10 +1,17 @@
 import { h, FunctionalComponent } from '@stencil/core';
+import { Button } from '../../../lib/ui/button/index';
 import { IconButton } from '../../../lib/ui/icon-button/index';
+import { Tooltip } from '../../../lib/ui/tooltip';
 import {
   PlusIcon,
   UploadIcon,
   DownloadIcon,
   SearchIcon,
+  BarChart3Icon,
+  PlayIcon,
+  SaveIcon,
+  SlidersIcon,
+  SpinnerIcon,
 } from '../../../lib/ui/icons/icons';
 import type { StatusFilter } from './test-case-filtering';
 
@@ -15,11 +22,20 @@ export interface TestCasesToolbarProps {
   passedCount: number;
   activeFilter: StatusFilter;
   isExportingTestSuite: boolean;
+  isExportingTestResults: boolean;
+  isRunningAll: boolean;
+  canRunAny: boolean;
+  useSave?: boolean;
+  isSaving?: boolean;
+  usePromptEditor?: boolean;
   searchQuery: string;
   isSearchExpanded: boolean;
   onAddTestCase: () => void;
   onImport: (file: File) => void;
   onExportSuite: () => void;
+  onExportResults: () => void;
+  onRunAll: () => void;
+  onSave?: () => void;
   onFilterChange: (filter: StatusFilter) => void;
   onSearchQueryChange: (query: string) => void;
   onSearchExpandedChange: (expanded: boolean) => void;
@@ -39,11 +55,20 @@ export const TestCasesToolbar: FunctionalComponent<TestCasesToolbarProps> = ({
   passedCount,
   activeFilter,
   isExportingTestSuite,
+  isExportingTestResults,
+  isRunningAll,
+  canRunAny,
+  useSave = false,
+  isSaving = false,
+  usePromptEditor = false,
   searchQuery,
   isSearchExpanded,
   onAddTestCase,
   onImport,
   onExportSuite,
+  onExportResults,
+  onRunAll,
+  onSave,
   onFilterChange,
   onSearchQueryChange,
   onSearchExpandedChange,
@@ -92,31 +117,34 @@ export const TestCasesToolbar: FunctionalComponent<TestCasesToolbarProps> = ({
             onChange={handleFileChange}
             accept=".json,application/json"
           />
-          <IconButton
-            variant="outline"
-            class="test-cases-toolbar__icon-btn"
-            onClick={onAddTestCase}
-            title="Add question"
-          >
-            <PlusIcon />
-          </IconButton>
-          <IconButton
-            variant="outline"
-            class="test-cases-toolbar__icon-btn"
-            onClick={handleFileSelect}
-            title="Import test suite"
-          >
-            <UploadIcon />
-          </IconButton>
-          <IconButton
-            variant="outline"
-            class="test-cases-toolbar__icon-btn"
-            onClick={onExportSuite}
-            loading={isExportingTestSuite}
-            title="Export test suite"
-          >
-            <DownloadIcon />
-          </IconButton>
+          <Tooltip content="Add question">
+            <IconButton
+              variant="outline"
+              class="test-cases-toolbar__icon-btn"
+              onClick={onAddTestCase}
+            >
+              <PlusIcon />
+            </IconButton>
+          </Tooltip>
+          <Tooltip content="Import test suite">
+            <IconButton
+              variant="outline"
+              class="test-cases-toolbar__icon-btn"
+              onClick={handleFileSelect}
+            >
+              <UploadIcon />
+            </IconButton>
+          </Tooltip>
+          <Tooltip content="Export test suite">
+            <IconButton
+              variant="outline"
+              class="test-cases-toolbar__icon-btn"
+              onClick={onExportSuite}
+              loading={isExportingTestSuite}
+            >
+              <DownloadIcon />
+            </IconButton>
+          </Tooltip>
         </div>
         <div class="test-cases-toolbar__filters">
           {FILTERS.map(filter => {
@@ -140,35 +168,77 @@ export const TestCasesToolbar: FunctionalComponent<TestCasesToolbarProps> = ({
           })}
         </div>
       </div>
-      <div class="test-cases-toolbar__search">
-        {isSearchExpanded ? (
-          <input
-            type="text"
-            class="test-cases-toolbar__search-input"
-            placeholder="Search questions…"
-            value={searchQuery}
-            ref={el => el?.focus()}
-            onInput={e =>
-              onSearchQueryChange((e.target as HTMLInputElement).value)
-            }
-            onBlur={e => {
-              // Read the live DOM value, not the searchQuery prop — Stencil's async
-              // re-render can lag a fast clear-then-blur, leaving the input stuck open.
-              if (!(e.target as HTMLInputElement).value.trim()) {
-                onSearchExpandedChange(false);
+      <div class="test-cases-toolbar__right">
+        <div class="test-cases-toolbar__search">
+          {isSearchExpanded ? (
+            <input
+              type="text"
+              class="test-cases-toolbar__search-input"
+              placeholder="Search questions…"
+              value={searchQuery}
+              ref={el => el?.focus()}
+              onInput={e =>
+                onSearchQueryChange((e.target as HTMLInputElement).value)
               }
-            }}
-          />
-        ) : (
+              onBlur={e => {
+                // Read the live DOM value, not the searchQuery prop — Stencil's async
+                // re-render can lag a fast clear-then-blur, leaving the input stuck open.
+                if (!(e.target as HTMLInputElement).value.trim()) {
+                  onSearchExpandedChange(false);
+                }
+              }}
+            />
+          ) : (
+            <IconButton
+              variant="outline"
+              class="test-cases-toolbar__icon-btn"
+              onClick={() => onSearchExpandedChange(true)}
+              title="Search questions"
+            >
+              <SearchIcon />
+            </IconButton>
+          )}
+        </div>
+        {usePromptEditor && (
+          <Tooltip content="Prompt editor">
+            <IconButton variant="outline" class="test-cases-toolbar__icon-btn">
+              <SlidersIcon />
+            </IconButton>
+          </Tooltip>
+        )}
+        <Tooltip content="Create report">
           <IconButton
             variant="outline"
             class="test-cases-toolbar__icon-btn"
-            onClick={() => onSearchExpandedChange(true)}
-            title="Search questions"
+            onClick={onExportResults}
+            loading={isExportingTestResults}
           >
-            <SearchIcon />
+            <BarChart3Icon />
           </IconButton>
+        </Tooltip>
+        {useSave && (
+          <Tooltip content="Save">
+            <IconButton
+              variant="outline"
+              class="test-cases-toolbar__icon-btn"
+              onClick={onSave}
+              loading={isSaving}
+            >
+              <SaveIcon />
+            </IconButton>
+          </Tooltip>
         )}
+        <Button
+          aria-label="Run all"
+          variant="primary"
+          size="md"
+          onClick={onRunAll}
+          disabled={isRunningAll || !canRunAny}
+          loading={isRunningAll}
+          icon={isRunningAll ? <SpinnerIcon /> : <PlayIcon />}
+        >
+          {isRunningAll ? 'Running…' : 'Run all'}
+        </Button>
       </div>
     </div>
   );
