+++ src/lib/ui/icons/icons.tsx
@@ -47,16 +47,15 @@ export const DownloadIcon: FunctionalComponent<IconProps> = ({
   </svg>
 );
 
-/** Document with text lines — used for Export Test Results. */
-export const FileTextIcon: FunctionalComponent<IconProps> = ({
+/** Three ascending bars — used for Create report. */
+export const BarChart3Icon: FunctionalComponent<IconProps> = ({
   class: cls,
 }) => (
   <svg {...baseProps} class={cls}>
-    <path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z" />
-    <polyline points="14 2 14 8 20 8" />
-    <line x1="16" y1="13" x2="8" y2="13" />
-    <line x1="16" y1="17" x2="8" y2="17" />
-    <polyline points="10 9 9 9 8 9" />
+    <path d="M3 3v18h18" />
+    <path d="M18 17V9" />
+    <path d="M13 17V5" />
+    <path d="M8 17v-3" />
   </svg>
 );
